requirements.txt
streamlit>=1.32.0
Pillow>=10.3.0
openai>=1.31.0